requirements.txt
langchain
langchain-community
langchain-huggingface
sentence-transformers
gpt4all
chromadb
pypdf
flask
flask-cors
pandas
openpyxl
requests
python-dotenv
